analytics/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
httpx